the compiler module was getting too bulky, refactored it to split out specials and operators into their own sub modules.
